Refactor: Move SuplidoresListRenderer to its own class
